+++ platforms/nextjs/src/theme/index.ts
@@ -1,5 +1,14 @@
-import { webVariants, reactThemeModes } from '@shallot-ui/platform-react'
+import {
+  reactThemeVariants,
+  reactThemeModes,
+  reactThemeGlobals,
+} from '@shallot-ui/platform-react'
 import { makeTheme } from '@shallot-ui/core-theme'
 
-export const DEFAULT_NEXT_THEME = makeTheme({}, webVariants, reactThemeModes)
+export const DEFAULT_NEXT_THEME = makeTheme(
+  {},
+  reactThemeVariants,
+  reactThemeModes,
+  reactThemeGlobals,
+)
 export type NextTheme = typeof DEFAULT_NEXT_THEME