+++ platforms/react/src/theme/globals/index.ts
@@ -1,0 +1,9 @@
+import { getColor } from '@shallot-ui/core-mixins'
+import { ThemeGlobals } from '@shallot-ui/core-theme'
+
+const reactThemeGlobals: ThemeGlobals = {
+  backgroundColor: getColor('Shading', 50),
+  color: getColor('Shading', 950),
+}
+
+export default reactThemeGlobals