+++ platforms/react/src/theme/variants/index.ts
@@ -2,6 +2,6 @@ import { ThemeVariants } from '@shallot-ui/core-theme'
 
 import Text from './Text'
 
-const webVariants: ThemeVariants = { Text }
+const reactThemeVariants: ThemeVariants = { Text }
 
-export default webVariants
+export default reactThemeVariants