+++ components/Checkbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @shallot-ui/component-checkbox
 
+## 3.0.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/core-theme@2.0.0-beta.8
+  - @shallot-ui/core-mixins@2.0.0-beta.8
+  - @shallot-ui/core-utils@2.0.0-beta.8
+
 ## 3.0.0-beta.7
 
 ### Patch Changes