+++ core/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/core-theme
 
+## 2.0.0-beta.8
+
+### Patch Changes
+
+- improves use of globals
+
 ## 2.0.0-beta.7
 
 ### Patch Changes